+++ example/kotlin/one-consumer-one-provider/consumer/src/main/kotlin/br/com/zup/pact/consumer/service/impl/ClientServiceImpl.kt
@@ -1,10 +1,11 @@
-package br.com.zup.pact.consumer.service
+package br.com.zup.pact.consumer.service.impl
 
 import br.com.zup.pact.consumer.dto.BalanceDTO
 import br.com.zup.pact.consumer.dto.ClientDetailsDTO
 import br.com.zup.pact.consumer.exception.ClientNotFoundException
 import br.com.zup.pact.consumer.integration.account.service.AccountIntegrationService
 import br.com.zup.pact.consumer.repository.ClientRepository
+import br.com.zup.pact.consumer.service.ClientService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
